renamed config_data.py to config_table_structure.py
