+++ pylovo/SyngridDatabaseConstructor.py
@@ -8,7 +8,7 @@ import psycopg2.errors
 import sqlparse
 
 from pylovo.config_loader import *
-from pylovo.config_data import *
+from pylovo.config_table_structure import *
 from pylovo.pgReaderWriter import PgReaderWriter
 from raw_data.preprocessing_scripts.process_trafos import process_trafos, get_trafos_processed_3035_geojson_path, \
     fetch_trafos, RELATION_ID, EPSG, get_trafos_processed_geojson_path
